yolo5/requirements.txt
pyyaml
loguru
boto3
pymongo
requests
